requirements.txt
# Core dependencies
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
pillow>=8.0.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Transformers and PEFT
transformers>=4.30.0
peft>=0.4.0

# CLIP (OpenAI)
git+https://github.com/openai/CLIP.git
ftfy
regex

# FAISS for similarity search
faiss-gpu>=1.7.0
# Use faiss-cpu if GPU is not available:
# faiss-cpu>=1.7.0

# Hydra for configuration
hydra-core>=1.3.0
omegaconf>=2.3.0

# Image processing
kornia>=0.6.0
opencv-python>=4.5.0

# Progress bar
tqdm>=4.60.0

# Visualization (optional)
matplotlib>=3.5.0

# Einops for tensor operations
einops>=0.6.0

# Pandas for data handling
pandas>=1.3.0

# Additional utilities
einops>=0.6.0
timm>=0.9.0
munch>=2.5.0
